Add Python session initialization module
