+++ python/isetcam/__init__.py
@@ -3,7 +3,8 @@
 from .vc_constants import vc_constants
 from .vc_get_image_format import vc_get_image_format
 from .quanta2energy import quanta_to_energy
-from .ie_init import ie_init, ie_init_session
+from .ie_init import ie_init
+from .ie_init_session import ie_init_session
 
 # Expose subpackages that mirror the MATLAB modules. These are currently
 # placeholders for future development.
