+++ python/isetcam/ie_init.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import os
-from datetime import datetime
 from typing import Any, Dict
 import warnings
 
@@ -12,9 +11,8 @@ try:
 except Exception:  # pragma: no cover - matplotlib might not be installed
     plt = None  # type: ignore
 
-# Global preferences and session storage
-ISET_PREFS: Dict[str, Any] = {}
-VC_SESSION: Dict[str, Any] = {}
+# Import the session initialization routine and globals
+from .ie_init_session import ie_init_session, vcSESSION, ISET_PREFS
 
 
 def ie_init(clear: bool = False) -> Dict[str, Any]:
@@ -45,45 +43,12 @@ def ie_init(clear: bool = False) -> Dict[str, Any]:
     ISET_PREFS.clear()
     ISET_PREFS.update({
         "waitbar": 0,
-        "init_clear": int(clear),
-        "font_size": 12,
+        "initClear": int(clear),
+        "fontSize": 12,
     })
 
     # Clear previous session data
-    VC_SESSION.clear()
+    vcSESSION.clear()
 
     # Initialize the session structure
     return ie_init_session()
-
-
-def ie_init_session() -> Dict[str, Any]:
-    """Initialize the global session structure.
-
-    Returns
-    -------
-    Dict[str, Any]
-        The newly created session dictionary stored in ``VC_SESSION``.
-    """
-    global VC_SESSION
-
-    VC_SESSION = {
-        "NAME": f"iset-{datetime.now():%Y%m%dT%H%M%S}",
-        "DIR": os.getcwd(),
-        "SELECTED": {
-            "SCENE": None,
-            "OPTICALIMAGE": None,
-            "ISA": None,
-            "VCIMAGE": None,
-            "DISPLAY": None,
-            "GRAPHWIN": None,
-        },
-        "SCENE": [None],
-        "OPTICALIMAGE": [None],
-        "ISA": [None],
-        "VCIMAGE": [None],
-        "GRAPHWIN": [],
-        "DISPLAY": [None],
-        "initHelp": 0,
-        "GUI": {"waitbar": ISET_PREFS.get("waitbar", 0)},
-    }
-    return VC_SESSION
